requirements.txt
flask
spacy
google-generativeai
